chore(SCTO): rename the PoC DAG and delete the old one
